+++ src/main/java/codesquad/fineants/spring/api/kis/service/KisService.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.kis;
+package codesquad.fineants.spring.api.kis.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
